requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.118.0
h11==0.16.0
httptools==0.6.4
idna==3.10
email-validator==2.2.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
python-slugify==8.0.4
